+++ src/main.rs
@@ -1,3 +1,6 @@
 fn main() {
-    dioxus_desktop::launch(coven::App)
+    dioxus_desktop::launch_cfg(
+        coven::App,
+        dioxus_desktop::Config::new(), // .with_custom_head(r#"<link rel="stylesheet" href="styles.css">"#.to_string()),
+    )
 }
